+++ __SITE__/unanunscript/bureau/partial/projet/helpers.rb
@@ -7,11 +7,4 @@ class Site
       values: Unan::UUProjet::TYPES, selected: params[:selected]})
   end
 
-  def menu_partage_projet params = nil
-    Form.build_select({
-      id: 'projet_partage', name: 'projet[partage]',
-      values: Unan::UUProjet::SHARINGS, selected: params[:selected]
-    })
-  end
-
 end